refactor: consolidate dual render systems into unified architecture
• Extract shared setAttributes function to attributes.ts module
• Eliminate legacy createElementLegacy and duplicated setAttributes code
• Unify render paths - both renderSimple and render now use same createElement flow
• Handle top-level arrays without fragment wrapper for backward compatibility
• Update test to be flexible about DOM attribute order
• Maintain full API backward compatibility while simplifying codebase architecture
